+++ tests/tclosures.nim
@@ -42,11 +42,11 @@ suite "agent closure slots":
       base = 100
 
     let
-      c1: proc(a: int) = proc(a: int) =
+      c1: proc(a: int) {.closure.} = proc(a: int) {.closure.} =
         base = a
       e = c1.rawEnv()
       p = c1.rawProc()
-      cc = ClosureRunner[proc(a: int)](rawEnv: e, rawProc: p)
+      cc = ClosureRunner[proc(a: int) {.nimcall.}](rawEnv: e, rawProc: p)
     connect(a, valueChanged, cc, ClosureRunner[proc(a: int) {.nimcall.}].callClosure)
 
     a.setValue(42)